chore: update dependencies and refactor Storybook configuration
- Removed deprecated dependencies and updated Storybook addons to version 9.0.1 across multiple package.json files.
- Refactored imports in Storybook stories to use '@storybook/react-vite' for improved compatibility.
- Updated routing imports from 'react-router-dom' to 'react-router' for consistency across components.
- Cleaned up external dependencies in Vite configuration for better performance.
- Enhanced example components to align with the latest routing practices.
