+++ apps/docs/package.json
@@ -15,23 +15,18 @@
     "@hookform/resolvers": "^3.9.1",
     "@lambdacurry/forms": "*",
     "@lambdacurry/medusa-forms": "*",
-    "@storybook/addon-essentials": "^8.6.7",
-    "@storybook/addon-interactions": "^8.6.7",
-    "@storybook/addon-links": "^8.6.7",
-    "@storybook/blocks": "^8.6.7",
-    "@storybook/react": "^8.6.7",
-    "@storybook/react-vite": "^8.6.7",
-    "@storybook/test": "^8.6.7",
+    "@storybook/addon-links": "^9.0.1",
+    "@storybook/react-vite": "^9.0.1",
     "react": "^19.0.0",
     "react-hook-form": "^7.51.0",
-    "react-router": "^7.0.0",
-    "react-router-dom": "^7.0.0",
+    "react-router": "^7.6.1",
     "remix-hook-form": "^7.0.1",
-    "storybook": "^8.6.7"
+    "storybook": "^9.0.1"
   },
   "devDependencies": {
     "@medusajs/ui-preset": "^2.8.3",
     "@react-router/dev": "^7.0.0",
+    "@storybook/addon-docs": "^9.0.1",
     "@storybook/test-runner": "^0.22.0",
     "@storybook/testing-library": "^0.2.2",
     "@tailwindcss/postcss": "^4.1.8",
