+++ apps/docs/src/examples/root-example.tsx
@@ -1,6 +1,6 @@
+import { Outlet } from 'react-router';
 // Example of setting up the middleware in root.tsx
 import { unstable_extractFormDataMiddleware } from 'remix-hook-form/middleware';
-import { Outlet } from 'react-router-dom';
 
 // Export the middleware for React Router 7
 export const unstable_middleware = [unstable_extractFormDataMiddleware()];
