+++ package.json
@@ -24,8 +24,7 @@
     "turbo": "^2.3.3"
   },
   "dependencies": {
-    "@changesets/cli": "^2.27.11",
-    "@medusajs/ui-preset": "^2.8.3"
+    "@changesets/cli": "^2.27.11"
   },
   "packageManager": "yarn@4.9.1"
 }
