+++ packages/components/package.json
@@ -64,8 +64,7 @@
     "next-themes": "^0.4.4",
     "react-day-picker": "8.10.1",
     "react-hook-form": "^7.53.1",
-    "react-router": "^7.0.0",
-    "react-router-dom": "^7.0.0",
+    "react-router": "^7.6.1",
     "remix-hook-form": "7.0.1",
     "sonner": "^1.7.1",
     "tailwind-merge": "^2.5.5",
