+++ packages/components/src/remix-hook-form/use-data-table-url-state.ts
@@ -1,18 +1,18 @@
 import { useCallback } from 'react';
-import { useSearchParams } from 'react-router-dom';
+import { useSearchParams } from 'react-router';
 import { type DataTableRouterState, dataTableRouterParsers } from './data-table-router-parsers';
 
 /**
  * Custom hook for managing URL state in data tables
- * 
+ *
  * This hook provides a clean interface for working with URL search parameters
  * in data table components, replacing the functionality previously provided by nuqs.
- * 
+ *
  * @returns An object containing the current URL state and a function to update it
  */
 export function useDataTableUrlState() {
   const [searchParams, setSearchParams] = useSearchParams();
-  
+
   // Parse URL search parameters using our custom parsers
   const urlState: DataTableRouterState = {
     search: dataTableRouterParsers.search.parse(searchParams.get('search')),
@@ -24,6 +24,7 @@ export function useDataTableUrlState() {
   };
 
   // Function to update URL search parameters
+  // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation>
   const setUrlState = useCallback(
     (newState: Partial<DataTableRouterState>) => {
       const updatedState = { ...urlState, ...newState };
@@ -63,7 +64,7 @@ export function useDataTableUrlState() {
       // Update the URL with the new search parameters
       setSearchParams(newParams, { replace: true });
     },
-    [urlState, setSearchParams]
+    [setSearchParams],
   );
 
   // Return the current URL state and the function to update it
@@ -72,7 +73,7 @@ export function useDataTableUrlState() {
 
 /**
  * Get the default state values for a data table
- * 
+ *
  * @param defaultStateValues Optional custom default values to override the standard ones
  * @returns A DataTableRouterState object with default values
  */
